safevoice/backend/agent/voice_verify.py
"""
SafeVoice AI — Voice Verification
Biometric matching + stress detection using Vertex AI.
Runs before any trigger is confirmed.
"""

import logging
import numpy as np
from typing import Optional
import vertexai
from vertexai.language_models import TextEmbeddingModel

from ..services.firestore_service import FirestoreService

logger = logging.getLogger(__name__)
firestore_svc = FirestoreService()


class VoiceVerifier:
    """
    Two-layer verification before any emergency trigger:
    1. Voice biometrics — is this the registered user?
    2. Stress detection — is there genuine distress in the voice?
    """

    BIOMETRIC_THRESHOLD = 0.82   # Cosine similarity — below this = reject
    STRESS_THRESHOLD    = 0.60   # Stress score 0-1 — below this = reject

    def __init__(self, user_id: str):
        self.user_id = user_id
        self._baseline_embedding: Optional[np.ndarray] = None

    async def load_baseline(self):
        """Load the user's registered voice baseline from Firestore."""
        data = await firestore_svc.get_user_voice_baseline(self.user_id)
        if data and "embedding" in data:
            self._baseline_embedding = np.array(data["embedding"])
            logger.info(f"[{self.user_id}] Voice baseline loaded")
        else:
            logger.warning(f"[{self.user_id}] No voice baseline found")

    async def verify(self, audio_chunk: bytes) -> dict:
        """
        Verify that the audio matches the registered user AND shows stress.

        Args:
            audio_chunk: Raw PCM audio bytes of the "Help Me" utterance

        Returns:
            dict with: verified (bool), biometric_score, stress_score, reason
        """
        if self._baseline_embedding is None:
            await self.load_baseline()

        # Run biometric check and stress detection in parallel
        biometric_score, stress_score = await asyncio.gather(
            self._check_biometrics(audio_chunk),
            self._detect_stress(audio_chunk),
        )

        verified = (
            biometric_score >= self.BIOMETRIC_THRESHOLD
            and stress_score >= self.STRESS_THRESHOLD
        )

        reason = None
        if not verified:
            if biometric_score < self.BIOMETRIC_THRESHOLD:
                reason = f"Voice mismatch (score: {biometric_score:.2f})"
            elif stress_score < self.STRESS_THRESHOLD:
                reason = f"No stress detected (score: {stress_score:.2f})"

        logger.info(
            f"[{self.user_id}] Verification: biometric={biometric_score:.2f}, "
            f"stress={stress_score:.2f}, verified={verified}"
        )

        return {
            "verified":        verified,
            "biometric_score": round(biometric_score, 3),
            "stress_score":    round(stress_score, 3),
            "reason":          reason,
        }

    async def _check_biometrics(self, audio_chunk: bytes) -> float:
        """
        Extract voice features and compare against registered baseline.
        Uses MFCC (Mel-Frequency Cepstral Coefficients) for speaker ID.

        In production: swap with a fine-tuned Vertex AI model.
        """
        try:
            features = self._extract_mfcc(audio_chunk)
            current_embedding = np.array(features)

            if self._baseline_embedding is None:
                return 0.0

            # Cosine similarity
            similarity = np.dot(current_embedding, self._baseline_embedding) / (
                np.linalg.norm(current_embedding) * np.linalg.norm(self._baseline_embedding)
                + 1e-8
            )
            return float(np.clip(similarity, 0.0, 1.0))

        except Exception as e:
            logger.error(f"[{self.user_id}] Biometric check failed: {e}")
            return 0.0

    async def _detect_stress(self, audio_chunk: bytes) -> float:
        """
        Detect stress patterns in voice using acoustic features:
        - Pitch variation (high stress = unstable pitch)
        - Speech rate (high stress = faster or irregular)
        - Energy level (high stress = higher RMS energy)
        - Formant analysis (stress alters vocal tract shape)

        Returns a stress score from 0.0 (calm) to 1.0 (high stress).
        """
        try:
            audio_array = np.frombuffer(audio_chunk, dtype=np.int16).astype(np.float32)
            audio_array /= 32768.0  # Normalize to [-1, 1]

            # RMS energy — high stress = louder voice
            rms = float(np.sqrt(np.mean(audio_array ** 2)))
            energy_score = min(rms * 5.0, 1.0)

            # Zero-crossing rate — high stress = more irregular
            zcr = float(np.mean(np.abs(np.diff(np.sign(audio_array)))) / 2)
            zcr_score = min(zcr * 3.0, 1.0)

            # Short-term variance — stress causes irregular amplitude
            frame_size = 512
            frames = [
                audio_array[i:i+frame_size]
                for i in range(0, len(audio_array) - frame_size, frame_size)
            ]
            if frames:
                frame_energies = [np.sqrt(np.mean(f**2)) for f in frames]
                variance_score = min(float(np.std(frame_energies)) * 10.0, 1.0)
            else:
                variance_score = 0.0

            # Weighted composite stress score
            stress_score = (
                energy_score   * 0.40 +
                zcr_score      * 0.30 +
                variance_score * 0.30
            )

            return float(np.clip(stress_score, 0.0, 1.0))

        except Exception as e:
            logger.error(f"[{self.user_id}] Stress detection failed: {e}")
            # On failure, allow trigger (safety > false negatives)
            return 1.0

    def _extract_mfcc(self, audio_chunk: bytes, n_mfcc: int = 40) -> list:
        """
        Simplified MFCC extraction.
        In production: use librosa.feature.mfcc() or a Vertex AI endpoint.
        """
        audio_array = np.frombuffer(audio_chunk, dtype=np.int16).astype(np.float32)
        audio_array /= 32768.0

        # Frame the signal
        frame_size = 512
        hop_size   = 256
        frames = np.array([
            audio_array[i:i+frame_size]
            for i in range(0, len(audio_array) - frame_size, hop_size)
        ])

        if len(frames) == 0:
            return [0.0] * n_mfcc

        # Apply Hamming window
        window = np.hamming(frame_size)
        windowed = frames * window

        # Power spectrum
        power = np.abs(np.fft.rfft(windowed, n=512)) ** 2

        # Mean across frames as simple feature vector (production: use full MFCC)
        features = np.mean(power, axis=0)[:n_mfcc]
        return features.tolist()


# Must be imported after class definition
import asyncio
