safevoice/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
google-genai==1.0.0
google-cloud-aiplatform==1.65.0
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.0
google-cloud-secret-manager==2.20.0
google-adk==0.5.0
twilio==9.3.0
httpx==0.27.0
numpy==1.26.4
pydantic==2.9.0
python-multipart==0.0.12
